+++ extensions/penpot/src/utils.ts
@@ -1,0 +1,4 @@
+export function iconAssetUri(thumbnailId: string) {
+  const endpoint = "https://design.penpot.app/";
+  return `${endpoint}assets/by-id/${thumbnailId}`;
+}