+++ extensions/penpot/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Windows] - {PR_MERGE_DATE}
 
 - Add Windows support
+- Add Icons for files
 
 ## [Initial Version] - 2024-11-17